+++ packages/cache/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@5.0.10...@aws-amplify/cache@5.0.11) (2023-01-19)
+
+**Note:** Version bump only for package @aws-amplify/cache
+
+
+
+
+
+## [5.0.10](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@5.0.9...@aws-amplify/cache@5.0.10) (2023-01-13)
+
+**Note:** Version bump only for package @aws-amplify/cache
+
+
+
+
+
+## [5.0.9](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@5.0.8...@aws-amplify/cache@5.0.9) (2023-01-10)
+
+**Note:** Version bump only for package @aws-amplify/cache
+
+
+
+
+
 ## [5.0.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@5.0.7...@aws-amplify/cache@5.0.8) (2022-12-27)
 
 **Note:** Version bump only for package @aws-amplify/cache